requirements.txt
Flask==2.0.1
boto3==1.18.46
pycryptodome==3.10.4
Werkzeug==2.0
python-dotenv==0.19.1
APScheduler==3.8.1
requests==2.26.0
pytz==2024.1